Reorganize and rebase
